rename 'test-demos' to 'test-contractHost', split out test-handoff into own file, and move 'test-gallery'. Remove tests for pixel-demo (subsumed by gallery)
